requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-ckeditor==6.0.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-resized==0.3.11
pilkit==2.0
Pillow==8.1.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
